requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
et-xmlfile==1.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pandas==1.5.2
psycopg2==2.9.5
pyTelegramBotAPI==4.8.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
xlrd==2.0.1
